Move transltion to own module
