+++ Browser/entry/__main__.py
@@ -1,6 +1,4 @@
 import contextlib
-import hashlib
-import inspect
 import json
 import logging
 import os
@@ -15,10 +13,12 @@ from pathlib import Path
 from typing import TYPE_CHECKING, List, Optional
 
 import click
-from robot import version as rf_version  # type: ignore
+from robot import version as rf_version
+
+from .translation import compare_translatoin, get_library_translaton
 
 if TYPE_CHECKING:
-    from .browser import Browser
+    from ..browser import Browser
 
 INSTALLATION_DIR = Path(__file__).parent / "wrapper"
 NODE_MODULES = INSTALLATION_DIR / "node_modules"
@@ -26,15 +26,10 @@ TIDY_TRANSFORMER_DIR = Path(__file__).parent / "tidy_transformer"
 # This is required because weirdly windows doesn't have `npm` in PATH without shell=True.
 # But shell=True breaks our linux CI
 SHELL = bool(platform.platform().startswith("Windows"))
-CURRENT_FOLDER = Path(__file__).resolve().parent
+ROOT_FOLDER = Path(__file__).resolve().parent.parent
 log_file = "rfbrowser.log"
-INSTALL_LOG = CURRENT_FOLDER / log_file
+INSTALL_LOG = ROOT_FOLDER / log_file
 PLAYWRIGHT_BROWSERS_PATH = "PLAYWRIGHT_BROWSERS_PATH"
-KEYWORD_NAME = "Keyword name"
-DOC_CHANGED = "Documentation update needed"
-NO_LIB_KEYWORD = "Keyword not found from library"
-MISSING_TRANSLATION = "Keyword is missing translation"
-MISSING_CHECKSUM = "Keyword tranlsaton is missing checksum"
 try:
     INSTALL_LOG.touch(exist_ok=True)
 except Exception as error:
@@ -224,7 +219,7 @@ def print_version(ctx, param, value):
     if not value or ctx.resilient_parsing:
         return
     _write_marker()
-    version_file = CURRENT_FOLDER / "version.py"
+    version_file = ROOT_FOLDER / "version.py"
     version_text = version_file.read_text()
     match = re.search(r"\"\d+\.\d+.\d+\"", version_text)
     browser_lib_version = match.group(0) if match else "unknown"
@@ -466,7 +461,7 @@ def launch_browser_server(browser, options):
     \b
     Example: re.search(r"ws://.*", console_output).group()
     """
-    from .browser import Browser, SupportedBrowsers
+    from ..browser import Browser, SupportedBrowsers
 
     logging.getLogger().setLevel(logging.INFO)
 
@@ -552,115 +547,6 @@ def transform(path: Path, wait_until_network_is_idle: bool):
     subprocess.run(cmd, check=False)
 
 
-def _get_library_translaton(
-    plugings: Optional[str] = None, jsextension: Optional[str] = None
-) -> dict:
-    from Browser import Browser
-
-    browser = Browser(plugins=plugings, jsextension=jsextension)
-    translation = {}
-    for function in browser.attributes.values():
-        translation[function.__name__] = {
-            "name": function.__name__,
-            "doc": function.__doc__,
-            "sha256": hashlib.sha256(function.__doc__.encode("utf-16")).hexdigest(),
-        }
-    translation["__init__"] = {
-        "name": "__init__",
-        "doc": inspect.getdoc(browser),
-        "sha256": hashlib.sha256(inspect.getdoc(browser).encode("utf-16")).hexdigest(),  # type: ignore
-    }
-    translation["__intro__"] = {
-        "name": "__intro__",
-        "doc": browser.__doc__,
-        "sha256": hashlib.sha256(browser.__doc__.encode("utf-16")).hexdigest(),  # type: ignore
-    }
-    return translation
-
-
-def _max_kw_name_lenght(project_tanslation: dict) -> int:
-    max_lenght = 0
-    for keyword_data in project_tanslation.values():
-        if (current_kw_length := len(keyword_data["name"])) > max_lenght:
-            max_lenght = current_kw_length
-    return max_lenght
-
-
-def _max_reason_lenght() -> int:
-    return max(
-        len(DOC_CHANGED),
-        len(NO_LIB_KEYWORD),
-        len(MISSING_TRANSLATION),
-        len(MISSING_CHECKSUM),
-    )
-
-
-def _get_heading(max_kw_lenght: int) -> List[str]:
-    heading = f"| {KEYWORD_NAME} "
-    next_line = f"| {'-' * len(KEYWORD_NAME)}"
-    if (padding := max_kw_lenght - len(heading) - 1) > 0:
-        heading = f"{heading}{' ' * padding}"
-        next_line = f"{next_line}{'-' * padding}"
-    reason = "Reason"
-    reason_padding = _max_reason_lenght() - len(reason)
-    heading = f"{heading}| {reason}{' ' * reason_padding}|"
-    next_line = f"{next_line} | {'-' * (_max_reason_lenght() -1)} |"
-    return [heading, next_line]
-
-
-def _table_doc_updated(lib_kw: str, max_name_lenght: int, reason: str) -> str:
-    line = f"| {lib_kw} "
-    if (padding := max_name_lenght - len(lib_kw) - 4) > 0:
-        line = f"{line}{' ' * padding}| {reason} "
-    else:
-        line = f"{line}| {reason} "
-    if reason == DOC_CHANGED:
-        line = f"{line}{' ' * 11}"
-    return f"{line}|"
-
-
-def _log_translation_table(table_body: List[str], heading: List[str]):
-    logging.info(
-        "Found differences between translation and library, see below for details."
-    )
-    for line_line in heading:
-        logging.info(line_line)
-    for line in table_body:
-        logging.info(line)
-
-
-def _compare(filename: Path, library_translation: dict):
-    with filename.open("r") as file:
-        project_translation = json.load(file)
-    max_kw_lenght = _max_kw_name_lenght(library_translation)
-    table_body = []
-    for lib_kw, lib_kw_data in library_translation.items():
-        project_kw_data = project_translation.get(lib_kw)
-        if not project_kw_data:
-            table_body.append(
-                _table_doc_updated(lib_kw, max_kw_lenght, MISSING_TRANSLATION)
-            )
-            continue
-        sha256_value = project_kw_data.get("sha256")
-        if not sha256_value:
-            table_body.append(
-                _table_doc_updated(lib_kw, max_kw_lenght, MISSING_CHECKSUM)
-            )
-            continue
-        if project_kw_data["sha256"] != lib_kw_data["sha256"]:
-            table_body.append(_table_doc_updated(lib_kw, max_kw_lenght, DOC_CHANGED))
-    for project_kw in project_translation:
-        if project_kw not in library_translation:
-            table_body.append(
-                _table_doc_updated(project_kw, max_kw_lenght, NO_LIB_KEYWORD)
-            )
-    if not table_body:
-        logging.info("Translation is valid, no updated needed.")
-    else:
-        heading = _get_heading(_max_kw_name_lenght(project_translation))
-        _log_translation_table(table_body, heading)
-
-
 @cli.command()
 @click.argument(
     "filename",
@@ -715,9 +601,16 @@ def translation(
     of keywords which documentation sha256 does not match. This will ease
     translation projects to identify keywords which documentation needs updating.
     """
-    translation = _get_library_translaton(plugings, jsextension)
+    translation = get_library_translaton(plugings, jsextension)
     if compare:
-        _compare(filename, translation)
+        if table := compare_translatoin(filename, translation):
+            logging.info(
+                "Found differences between translation and library, see below for details."
+            )
+            for line in table:
+                logging.info(line)
+        else:
+            logging.info("Translation is valid, no updated needed.")
     else:
         with filename.open("w") as file:
             json.dump(translation, file, indent=4)
